Reorganize testing config
