--- scripts/scraper-ng/test/jest.config.js
@@ -1,3 +1,0 @@
-module.exports = {
-  setupFilesAfterEnv: ["./jest.setup.js"],
-};
